full project mern_app CRUD post

+++ back-end/controllers/goalController.js
@@ -6,12 +6,7 @@ const Goal = require('../models/goalModel')
 //@route GET: api/goals 
 const getGoal = asyncHandler(async (req,res) => {
 
-    let goals
-    if(req.params.id)
-    {
-        goals = await Goal.findById(req.params.id)
-    }
-    else goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
 
     res.status(200).json(goals)
 })
@@ -19,7 +14,6 @@ const getGoal = asyncHandler(async (req,res) => {
 //@desc Post goals
 //@route POST: api/goals
 const postGoal = asyncHandler(async (req,res) => {
-
     if(!req.body?.text)
     {
         res.status(400)
@@ -27,45 +21,81 @@ const postGoal = asyncHandler(async (req,res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
-    res.status(200).json({
+
+    res.status(200).json(
         goal
-    })
+    )
 })
 
 //@desc Put goals
 //@route PUT: api/goals/:id
 
 const putGoal = asyncHandler(async (req,res) => {
-    const goal  = await Goal.findById(req.params.id)
+    console.log(req.body)
+    const goal  = await Goal.findById(req.body.id)
     if(!goal)
     {
         res.status(400)
         
         throw new Error('Goal not found!')
     }
+    if(!req.user)
+    {
+        res.status(401)
+        throw new Error('User not found!')
+    }
 
+    if(goal.user.toString() !== req.user.id)
+    {
+        res.status(401)
+        throw new Error('User not authorized!')
+    }
     // new : true là trả về document sau khi update
-    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
+    const updatedGoal = await Goal.findByIdAndUpdate(req.body.id, {text: req.body.text}, {
         new: true,
     })
-    res.status(200).json({
+
+    res.status(200).json(
         updatedGoal
-    })
+    )
 })
 
 //@desc Delete goals
 //@route DELETE: api/goals/:id
 const deleteGoal = asyncHandler(async (req,res) => {
 
-    const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
+    const goal = await Goal.findById(req.params.id)
 
-    if (!deletedGoal) {
+    if (!goal) {
         return res.status(404).json({ message: 'Goal not found' })
     }
 
-    res.status(200).json({ message: 'Goal deleted successfully' })
+    if(!req.user)
+    {
+        res.status(401)
+        throw new Error('User not found!')
+    }
+
+    if(goal.user.toString() !== req.user.id)
+    {
+        res.status(401)
+        throw new Error('User not authorized!')
+    }
+
+    const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
+    if(!deletedGoal)
+    {
+        res.status(401)
+        throw new Error('Goal delete failed!')
+    }
+
+    res.status(200).json({ 
+        message: 'Goal deleted successfully',
+        id: req.params.id
+     })
 })
 
 module.exports = {
